+++ scripts/crispy-critters.js
@@ -46,76 +46,118 @@ const THEME_DEFAULT = {
   visibility: "visible"
 };
 
+/*
+SERIES:
+[<theme ID/texture webp file name>, <theme display name>, <accent color>, <optional overrides>]
+
+Default theme settings:
+- font: Montserrat
+- foreground (numbering color): white
+- background (dice color): <accent color>
+- outline: <accent color>
+- edge: <accent color>
+- material: plastic
+- visibility: visible
+
+You can override all default settings when defining a theme
+*/
+
 const SERIES = {
   "Astral Series": [
-    ["black-hole", "Black Hole", "#fde5ce"],
-    ["cataclysm", "Cataclysm", COLOR.BLACK],
-    ["cosmic-chaos", "Cosmic Chaos", "#fcae31"],
-    ["ember", "Ember", "#6d186c"],
-    ["moon", "Moon", "#32333e"],
-    ["nebula", "Nebula", "#003b99"],
-    ["radiance", "Radiance", "#efb10c"],
-    ["solaris", "Solaris", "#f64933"],
-    ["space-oddity", "Space Oddity", "#1d4e6d"],
-    ["tempest", "Tempest", "#066f73"],
-    ["tide", "Tide", "#358363"]
+    ["as-black-hole", "Black Hole", "#fde5ce"],
+    ["as-cataclysm", "Cataclysm", COLOR.BLACK],
+    ["as-cosmic-chaos", "Cosmic Chaos", "#fcae31"],
+    ["as-ember", "Ember", "#6d186c"],
+    ["as-moon", "Moon", "#32333e"],
+    ["as-nebula", "Nebula", "#003b99"],
+    ["as-radiance", "Radiance", "#efb10c"],
+    ["as-solaris", "Solaris", "#f64933"],
+    ["as-space-oddity", "Space Oddity", "#1d4e6d"],
+    ["as-tempest", "Tempest", "#066f73"],
+    ["as-tide", "Tide", "#358363"]
   ],
 
   "Candy Series": [
-    ["candy-hearts", "Candy Hearts", "#bb0306"],
-    ["cotton-candy", "Cotton Candy", "#fb87c3"],
-    ["mint-chocolate-chip", "Mint Chocolate Chip", "#7fa96e", BLACK_TEXT],
-    ["saltwater-taffy", "Saltwater Taffy", "#7bcbe2", BLACK_TEXT],
-    ["sweets", "Sweets", "#fa9633", BLACK_TEXT],
-    ["waffles", "Waffles", "#f3a936"]
+    ["cs-candy-hearts", "Candy Hearts", "#bb0306"],
+    ["cs-cotton-candy", "Cotton Candy", "#fb87c3"],
+    ["cs-mint-chocolate-chip", "Mint Chocolate Chip", "#7fa96e", BLACK_TEXT],
+    ["cs-saltwater-taffy", "Saltwater Taffy", "#7bcbe2", BLACK_TEXT],
+    ["cs-sweets", "Sweets", "#fa9633", BLACK_TEXT],
+    ["cs-waffles", "Waffles", "#f3a936"]
+  ],
+
+  "Crystal Series": [
+    ["crs-amethyst", "Amethyst", "#9966cc"],
+    ["crs-diamond", "Diamond", "#eaf6ff", BLACK_TEXT],
+    ["crs-emerald", "Emerald", "#50c878"],
+    ["crs-ruby", "Ruby", "#e0115f"],
+    ["crs-sapphire", "Sapphire", "#0f52ba"],
+    ["crs-topaz", "Topaz", "#ffc857"]
+  ],
+
+  "Mythic Marble Series": [
+    ["mms-blazewave", "Blazewave", "#fd7c01"],
+    ["mms-dreamweaver", "Dreamweaver", "#6232cb"],
+    ["mms-enchanted-forest", "Enchanted Forest", "#649e63"],
+    ["mms-funhouse-fever", "Funhouse Fever", "#fcc54a"],
+    ["mms-funkalicious", "Funkalicious", "#de537d"],
+    ["mms-necrosis", "Necrosis", "#554085"],
+    ["mms-nefarious-knight", "Nefarious Knight", "#780203"],
+    ["mms-neptunes-wake", "Neptune's Wake", "#aad4d2"],
+    ["mms-phoenix", "Phoenix", "#fd5b02"],
+    ["mms-skybreaker", "Skybreaker", "#3e5979"],
+    ["mms-toxic-twirls", "Toxic Twirls", "#cbf50a"]
   ],
 
   "Obsidian Series": [
-    ["ice", "Ice", "#80a6ff"],
-    ["magma", "Magma", "#c73032"],
-    ["ooze", "Ooze", "#9cd604"]
+    ["os-ice", "Ice", "#80a6ff"],
+    ["os-magma", "Magma", "#c73032"],
+    ["os-ooze", "Ooze", "#9cd604"]
   ],
 
   "Pixie Series": [
-    ["pixie-blue", "Pixie Blue", "#2badff"],
-    ["pixie-pink", "Pixie Pink", "#ff89ff"]
+    ["ps-pixie-blue", "Pixie Blue", "#2badff"],
+    ["ps-pixie-pink", "Pixie Pink", "#ff89ff"]
   ],
 
   "Signature Series": [
-    ["barbarian", "Barbarian", "#e7623e"],
-    ["bard", "Bard", "#ab6dac"],
-    ["cleric", "Cleric", "#91a1b2"],
-    ["druid", "Druid", "#7a853b"],
-    ["fighter", "Fighter", "#7f513e"],
-    ["monk", "Monk", "#51a5c5"],
-    ["paladin", "Paladin", "#b59e54"],
-    ["ranger", "Ranger", "#507f62"],
-    ["rogue", "Rogue", "#555752"],
-    ["sorcerer", "Sorcerer", "#992e2e"],
-    ["warlock", "Warlock", "#7b469b"],
-    ["wizard", "Wizard", "#2a50a1"]
+    ["ss-barbarian", "Barbarian", "#e7623e"],
+    ["ss-bard", "Bard", "#ab6dac"],
+    ["ss-cleric", "Cleric", "#91a1b2"],
+    ["ss-druid", "Druid", "#7a853b"],
+    ["ss-fighter", "Fighter", "#7f513e"],
+    ["ss-monk", "Monk", "#51a5c5"],
+    ["ss-paladin", "Paladin", "#b59e54"],
+    ["ss-ranger", "Ranger", "#507f62"],
+    ["ss-rogue", "Rogue", "#555752"],
+    ["ss-sorcerer", "Sorcerer", "#992e2e"],
+    ["ss-warlock", "Warlock", "#7b469b"],
+    ["ss-wizard", "Wizard", "#2a50a1"],   
   ],
 
   "Spooky Series": [
-    ["abyssal-grave", "Abyssal Grave", "#005255"],
-    ["dexters-delight", "Dexter's Delight", "#630201", BLACK_TEXT_WHITE_OUTLINE],
-    ["eldritch-night", "Eldritch Night", "#784a78"],
-    ["god-damaged", "God Damaged", "#af03ef"],
-    ["hypno-creep", "Hypno Creep", "#161616"],
-    ["party-skulls", "Party Skulls", "#07bfb0"],
-    ["spooky-ghosts", "Spooky Ghosts", "#1b1e2a"],
-    ["zombies", "Zombies", "#7b9c86"]
+    ["sps-abyssal-grave", "Abyssal Grave", "#005255"],
+    ["sps-bones", "Bones", "#593a14"],
+    ["sps-dexters-delight", "Dexter's Delight", "#630201", BLACK_TEXT_WHITE_OUTLINE],
+    ["sps-eldritch-night", "Eldritch Night", "#784a78"],
+    ["sps-god-damaged", "God Damaged", "#af03ef"],
+    ["sps-hypno-creep", "Hypno Creep", "#161616"],
+    ["sps-party-skulls", "Party Skulls", "#07bfb0"],
+    ["sps-spooky-ghosts", "Spooky Ghosts", "#1b1e2a"],
+    ["sps-wall-of-eyes", "Wall of Eyes", COLOR.BLACK],
+    ["sps-zombies", "Zombies", "#7b9c86"],
   ],
 
   "Wildcard Series": [
-    ["bbc", "BBC (Big Beautiful Cow)", COLOR.WHITE, BLACK_TEXT],
-    ["checkers", "Checkers", "#b7050a"],
-    ["dragon-scale", "Dragon Scale", "#cd4500"],
-    ["gabbys-grass", "Gabby's Grass", "#66a509"],
-    ["san-diego-sunset", "San Diego Sunset", "#ef0200"],
-    ["synthwave", "Synthwave", "#600095"],
-    ["vampira", "Vampira", "#c73032"],
-    ["winter-is-coming", "Winter Is Coming", "#80a6ff"]
+    ["ws-bbc", "BBC (Big Beautiful Cow)", COLOR.WHITE, BLACK_TEXT_WHITE_OUTLINE],
+    ["ws-checkers", "Checkers", "#b7050a"],
+    ["ws-dragon-scale", "Dragon Scale", "#cd4500"],
+    ["ws-fort-ardovon", "Fort Ardovon", "#f5c709"],
+    ["ws-gabbys-grass", "Gabby's Grass", "#66a509"],
+    ["ws-san-diego-sunset", "San Diego Sunset", "#ef0200"],
+    ["ws-synthwave", "Synthwave", "#600095"],
+    ["ws-vampira", "Vampira", "#c73032"],
+    ["ws-winter-is-coming", "Winter Is Coming", "#80a6ff"]
   ]
 };