+++ CHANGELOG.md
@@ -1,23 +1,52 @@
 # 🦗 Changelog
 
+## v1.3 - 5-19-26
+### 🔧 Changes
+- Fixed naming conflict of `Obsidian Series: Ice` with DSN's stock `Ice` theme
+- Renamed texture files to include a series indicator to prevent future naming conflicts
+
+### ✨ New Dice
+- Added `Crystal Series: Amethyst`
+- Added `Crystal Series: Diamond`
+- Added `Crystal Series: Emerald`
+- Added `Crystal Series: Ruby`
+- Added `Crystal Series: Sapphire`
+- Added `Crystal Series: Topaz`
+- Added `Mythic Marble Series: Blazewave`
+- Added `Mythic Marble Series: Dreamweaver`
+- Added `Mythic Marble Series: Enchanted Forest`
+- Added `Mythic Marble Series: Funhouse Fever`
+- Added `Mythic Marble Series: Funkalicious`
+- Added `Mythic Marble Series: Necrosis`
+- Added `Mythic Marble Series: Nefarious Knight`
+- Added `Mythic Marble Series: Neptune's Wake`
+- Added `Mythic Marble Series: Phoenix`
+- Added `Mythic Marble Series: Skybreaker`
+- Added `Mythic Marble Series: Toxic Twirls`
+- Added `Spooky Series: Bones`
+- Added `Spooky Series: Wall of Eyes`
+- Added `Wildcard Series: Fort Ardovon`
+
+---
+
 ## v1.2 — 5-12-26
 ### 🔧 Changes
 - Further code optimization
-- Changed Candy Series: Saltwater Taffy's default loadout to use black numbering
-- Changed Candy Series: Sweets' default loadout to use black numbering
+- Changed `Candy Series: Saltwater Taffy` default loadout to use black numbering
+- Changed `Candy Series: Sweets` default loadout to use black numbering
 
 ---
 
 ## v1.1 - 5-12-26
-### ✨ New
-- Added Astral Series: Black Hole
-- Added Astral Series: Cataclysm
-- Added Candy Series: Mint Chocolate Chip
-- Added Candy Series: Saltwater Taffy
-- Added Candy Series: Waffles
-- Added Spooky Series: Dexter's Delight
-- Added Wildcard Series: Checkers
-- Added Wildcard Series: Synthwave
+### ✨ New Dice
+- Added `Astral Series: Black Hole`
+- Added `Astral Series: Cataclysm`
+- Added `Candy Series: Mint Chocolate Chip`
+- Added `Candy Series: Saltwater Taffy`
+- Added `Candy Series: Waffles`
+- Added `Spooky Series: Dexter's Delight`
+- Added `Wildcard Series: Checkers`
+- Added `Wildcard Series: Synthwave`
 
 ## v1.0 - 5-11-26
 - Initial release